Fixed ExponentiallyDecayingSample by using nano-seconds ticks as in the Java verion, instead of either the Stopwatch ticks or datetime ticks.
Conflicts:
	src/metrics/Core/MeterMetric.cs

+++ src/metrics/Stats/UniformSample.cs
@@ -68,7 +68,7 @@ namespace metrics.Stats
             }
             else
             {
-                var random = Math.Abs(Support.Random.NextLong()) % count;
+                var random = Math.Abs(Support.Random.NextInt()) % count;
                 if (random < _values.Length)
                 {
                     var index = (int) random;
